refactor: rename package step2 for calculator
