+++ src/main/kotlin/calculator/InvalidCalculatorInput.kt
@@ -1,3 +1,3 @@
-package step2
+package calculator
 
 class InvalidCalculatorInput : IllegalArgumentException("입력값이 유효하지 않습니다.")
