+++ src/test/kotlin/calculator/StringCalculatorSpec.kt
@@ -1,8 +1,7 @@
-package step2
+package calculator
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
-import io.kotest.data.forAll
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
